Completely rewrite Control
This is a major cleanup, and prepares the core for later. It adopts a
structure that is arguably more complex but this is necessary for the
app to grow easily in the future and for maintainability.

The log in system is now ready, and is entirely powered by AngularJS.
Routing is also ready, and controllers can be lazy-loaded by routing.

There are still some issues to fix and some stuff to tweak and improve.
This commit also removes unnecessary stuff from the repo to keep it
clean! Please note that I might have forgotten some changes in this
commit message. If so, I'm sorry :)