Use npm and Bower for managing dependencies
I'm delegating the dependency management to npm and Bower:

- npm will be in charge of Bower, Grunt and other tools *we* will
  directly use when working on Control;

- Bower will take care of the front-end dependencies, such as jQuery,
  Bootstrap, etc.

This allows for a cleaner and simpler project in overall; and more
importantly, it is much simpler to manage the dependencies and update
them using npm and Bower!

Therefore, this commit adds in the required packages and the
dependencies installed by Bower, and updates index.html to point to the
new files.

Note: I have got rid of angular-cookies as we don't need it, and
replaced bootstrap-colorpicker by angular-bootstrap-colorpicker, which
is integrated to AngularJS and allows binding (via the scope).
So currently the colorpicker is broken, but this will be worked on in a
later commit. The (now) redundant deps have been removed.

This also introduces the concept of "building" Control (not really a
build though) with Grunt: every time we are deploying to production,
Grunt will be ran and will generate a production version; the app
itself, its assets and dependencies will get optimised (minified).

This allows for a faster loading time while still allowing the devs
to develop and debug as easily as before.

The production build will be located in dist/. Do not edit any files
there, as they will be overridden during the next build. Instead,
if you want to work on the project, edit the source, which is, as you
may have guessed, in src/.

So we need to have another branch for regular dev and testing, and
a branch for the production version:

- master will be the production branch.

  New versions that will be merged in from another branch need to be
  tested before the merge, as it will directly and automatically
  be built and be deployed to production.

- dev will be the dev branch. (Kind of obvious, isn't it?)

  You can do whatever you want in this branch.
  It will contain the production build as well, as we also need to test
  it too (although it's not really a "production" build in this case).

  Merge to master when you have made sure everything is fine.

  Feature branches can be merged to dev when they are ready but still
  need work and testing.