Add a config module
Add a config module named "config" that can be injected via DI
throughout the app. The config itself is dynamically generated by
grunt-ng-constant, from what's defined in Gruntfile.js.

This is useful for defining development and production env variables;
for example, there's now ENV.apiEndpoint that contains the server to
use in the app.

For the dev environment, it is the dev ITFrame server.
For the prod environment, it is itframe.shoutca.st.

This might also be useful for other stuff in the future.