Use the server defined in the config
Use the server defined in the config, so that in a dev env, the dev
server will handle the API calls, whereas the prod server
(itframe.shoutca.st) will be used in a production environment.

This commit also ditches passing the email and token manually on every
request, as this is now useless: requests already send an Authorization
header and this header is now read by ITFrame.