Update the authentication helpers
Update the authentication-related functions and helpers following a
server-side change in how authentication is handled.

Note that the email address isn't used nor stored anymore.

The AuthService.logout() method now only deletes the token locally
from the local storage; it does not send a logout request to the
server (as there is no such code to handle that server-side).

While this may seem insecure, the tokens' validity is very short:
it is only 30 minutes. This will be compensated by sending a
"keep alive" request every 25 minutes.