Send a keep-alive request every 25 minutes
Send a keep-alive request to the server every 25 minutes (5 minutes
before a token expires) to request a new token to use.

Note that the token should be invalidated and added to a blacklist once
it is invalidated. Protected routes should return 401 if the token is
expired. I'll probably implement this in a future commit to ITFrame.