Show alerts when something goes wrong server-side
Show nice modals when something goes wrong on the server to inform the
user that something's gone wrong (obviously) and tell them to try again
and potentially report it to us if it occurs again.

It currently handles 500 (Internal Server Error) and 503 (Service
Temporarily Unavailable) and displays the appropriate message and the
action the user needs to take (if any).