Rewrite RequestApp
This commit renames the folder requestapp to request-app, which in my
honest opinion looks better (at least to me). In the same way, I've
renamed RequestappCtrl and RequestappService to (respectively)
RequestAppCtrl and RequestAppService.

So, this is the first major change.

The second major change is the new route used to request apps. It is
now just for a particular service and the related account's username is
in the route. Therefore, manage.html has been updated to link to the
new URLs.

The controller is now loaded directly as a dependency required in
src/app/app.js. The service list is also injected *correctly* and more
importantly, there is no code duplicated! Ah, the beauty of code re-use
and dependency injection.

I've also completely rewritten RequestAppCtrl and RequestAppService.

RequestAppCtrl has been rewritten for the new route scheme, and is
objectively much cleaner. There are also critical checks to ensure the
m̶a̶l̶i̶c̶i̶o̶u̶s̶ user doesn't request apps for a service
that doesn't exist, that they don't own, that is not active or that is
otherwise not a valid service eligible for mobile apps (i.e.,
non-streaming servers and free servers). Those requests will be
rejected with an error message and redirected back to the manage page.
Of course, the server also needs to check for those cases. That'll
likely be implemented soon. At least, I hope so!

The code is now also much simpler and shorter, thanks to code re-use.
Same for RequestAppService which now simply has one and only one method
for sending app requests. Code isn't duplicated there anymore, too.

Another major change is the complete layout change and UX improvements.
It should be cleaner and look better, both on PC *and* on mobile
devices! I won't list all changes; that would be way too long!

PS: I've also fixed a typo in an API call.