WIP: Add #NowPlaying management
Add #NowPlaying management with full and automated support for all
settings, including Twitter token and secret which can be obtained
easily through a button.

This requires adding a PHP script that handles the Twitter OAuth part,
as secrets really shouldn't be stored on the client-side. However, the
script is pretty invisible from the end user, as all it does is
redirecting to the Twitter authorise page and then passing the data
back to the client.

It is NOT ready for production. See the FIXME comments in the source
code for more information, but here's a gist of what needs to be fixed:

- #NowPlaying description.

  I desperately need a description of what #NowPlaying is and what it
  does. (cc. @ethan and @aaron)

  As soon as I receive a complete description, I'll make a commit that
  completes this commit and merge this feature into production.

- And more importantly, the backend (ITFrame).

  This commit should serve as a reference for the implementation and
  the API that Control uses for this feature. Unless absolutely
  required, the implementation likely won't change, so no further
  changes are required for this issue.

*Until all of those issues are fixed, do NOT merge.*