Further improve UX by hiding secret
The secret is actually of no use to the user, and leaving it as is,
visible to everyone that can watch the screen, is nothing but a
security and UX issue: it is useless to the user, and could be stolen
by someone behind the user.

Instead, just leave the token displayed and show whether or not they
have authorised us to post, and if so, using which Twitter account.