WIP: Add the Cast sections and items
This commit adds a Cast section and related items to the sidebar menu.

For now, none of the items works, as they're all missing a controller
and a partial -- but at least they are ordered and have their routes,
labels and icons ready now.

Note: they are made visible for all services to make it easier to work
on it. Be sure to change it before merging into master!