Fix a lot of bugs caused by incorrect event handlers and events
This commit fixes a whole range of bugs, some easy to trigger, some
harder to find, caused by incorrect event handlers and missing events
in angular-route-segment.

src/app/common/route-checks.js is renamed to route-event-handlers.js as
it clarifies the content of that file.

This also fixes a bug which would cause route-specific page titles not
to show up.