Fix the Session.update() implementation
The update() method was destroying and recreating the whole session
(and broadcasting the associated events) on every update. This is a
bit too overkill.

So this commit changes the implementation and makes it simply update
the token property in the active session, in the local storage and
broadcasts the `sessionUpdated` event.

This also makes the test for Session pass.