Player: Only allow the user to upload the logo
If there's one thing we have learned from the app request submit and
edit form, it's that clients are STUPID and don't know how to properly
fill in a field that asks them for a direct link to an image file. They
end up providing incorrect links and even links to private FB photos.
They just CANNOT do this correctly. So, we have to hold their hands and
force them to select a file and upload it. We cannot trust them enough
to allow them to provide a link directly. <end of rant>

This commit also makes the logo optional.

See issue #1.