Enable full continuous delivery
After testing the master => production workflow for a while, it turns
out that we do not really need to control where the code gets pushed,
and having to merge to deploy proved to be a hassle (at least to me).

So this commit changes the CI config to just deploy any merged commits
on master automatically, which allows us to have automated, real
continuous delivery.