Switch to Webpack
This commit has everything needed to switch to Webpack and remove
grunt and bower from our workflow.

Since the Control project started, the JS ecosystem has changed a lot
and so has Control. Control has become a medium to large app and the
previous build system was becoming a pain: missed new files, updates,
manual <script> tags… and a fragile build process that could break
only in the production build or in the test builds.

Additionally, why use both bower and npm when they have similar roles?
npm now has most of the packages we used to rely on bower for, and for
the rare packages that are not available or kept up-to-date on the npm
registry, npm can also pull directly from a git repository.

So, the past days I have been working on replacing them with something
more modern, that allows Control's codebase to be more modular, easier
to maintain and easier to edit. And the solution for this is Webpack.

Note that I haven't had time to update the whole codebase to ES6;
and because of lack of time, some parts of Control that were edited
relatively later have basically just been ported without a lot of
changes, with the sole goal of making them work. A cleanup will come
later.