Revert to Angular 1.4.8
Formly has issues with 1.5.0 (which has breaking changes for $compile)
and cause the formly-powered forms not to work properly. As there is
nothing related to 1.5.0 (no issue, no mention) on formly's repo, this
commit reverts Angular to 1.4.8. We'll wait until 1.5 gets released and
all modules we use are compatible to update.