Reformat the code to 2 spaces, double quotes
This commit just changes code formatting in the whole codebase. This is
to be more consistent with other client-side projects like Player
and also server-side projects like ITFrame which use double quotes.

Since this is only a code formatting change, it has no consequences on
code behaviour and user interface.