NowPlaying: Make the doc link label more descriptive
Instead of "click here", it now says where it links to, and what
clicking on it does. This was among the earliest features implemented
in Control, which is why this sort of thing managed to be in our code.
(It's the only instance of "click here" links in Control.)

Fixes #10.