Apps: Block app request edits in some cases
This commit adds client-side code to block app request edits in the
following cases:

- When an app request is Pending (or Update Pending), and the submitted
  or last updated date is more than 5 minutes ago. This is to prevent
  users from updating their requests simply because they think that'll
  get their requests processed faster.

- When the request status is In Progress. It's obviously a bad idea to
  update an app request when it is in the middle of a submission or
  of a build for iOS apps.

- When the app request is for an iOS app, and the status is Submitted
  or Update Submitted. We cannot update apps when they've been
  submitted for review to Apple, so let's just deny those.

For the moment, the check merely disables the button. It doesn't really
prevent someone with sufficient knowledge from submitting an update
(unless we add a server-side check, but that's probably overkill,
considering that most (all?) of our users do not have the required
skills to even know it's possible).

Fixes issue #3.

PS: Parts of the .footer-links CSS is moved into a new shared class,
.flex-left-right to reuse more code or (more precisely) styles.