Fix unclosable, empty alert when session times out
Looks like alert.destroy() does not work as expected at all. So instead
we are now skipping destroying and showing another alert altogether if
an alert was already shown in the last 5 seconds. This does not
completely fix the issue (which was that multiple session timeout
alerts shouldn't be shown at the same time) but it reduces it a lot.