Apps: Add a button to convert into selfHosted
Easier to notice for clients, easier for them to convert their apps
into a selfHosted request, fewer clicks as well. Also not subject to
the edit restrictions, and allows having both the restrictions and
giving the user the right to edit their app request to turn it into a
selfHosted request.